requirements.txt
# Core dependencies
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-c>=0.20.0

# GitLab API
python-gitlab>=3.15.0
requests>=2.31.0

# Parsing and utilities
unidiff>=0.7.5  # For parsing git diffs
pyyaml>=6.0
click>=8.1.0    # CLI framework

# Metrics and analysis
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
